Change the OP_ListXX opcodes to use 64 bit integers (CVS 1372)
FossilOrigin-Name: da9b3dce334bde99360db45b4a3d41be519ec2d1

+++ src/vdbeInt.h
@@ -219,7 +219,7 @@ struct Keylist {
   int nUsed;        /* Next unwritten slot in aKey[] */
   int nRead;        /* Next unread slot in aKey[] */
   Keylist *pNext;   /* Next block of keys */
-  int aKey[1];      /* One or more keys.  Extra space allocated as needed */
+  i64 aKey[1];      /* One or more keys.  Extra space allocated as needed */
 };
 
 /*
